upload of the main project file
In this file the research of the project proposed in the course is done. An LLM model, called3, is achieved, which is given a well-structured prompt to obtain a specific result for each sentence. Performance and statistics are uploaded to COMET, a website where you can see the general statistics of our project.

In the end I have added more research points that I would like to do out of interest and on my own.